Handle authentification process

+++ src/Controller/PostController.php
@@ -15,15 +15,20 @@ class PostController extends AbstractController
     #[Route('/', name: "home")]
     public function index(ManagerRegistry $doctrine): Response
     {
+
         $repository = $doctrine->getRepository(Post::class);
         $posts = $repository->findAll();
         //SELECT ALL FROM 'post'
-        return $this->render('post/index.html.twig', ["posts" => $posts]);
+        return $this->render(
+            'post/index.html.twig',
+            ["posts" => $posts]
+        );
     }
 
     #[Route('/post/new')]
     public function create(Request $request, ManagerRegistry $doctrine): Response
     {
+        $this->denyAccessUnlessGranted('IS_AUTHENTICATED__FULLY');
         $post = new Post();
         $form = $this->createForm(PostType::class, $post);
         $form->handleRequest($request);
@@ -43,6 +48,7 @@ class PostController extends AbstractController
     #[Route('/post/delete/{id<\d+>}', name: 'delete-post')]
     public function delete(Post $post, ManagerRegistry $doctrine): Response
     {
+        $this->denyAccessUnlessGranted('IS_AUTHENTICATED__FULLY');
         $em = $doctrine->getManager();
         $em->remove($post);
         $em->flush();
@@ -53,6 +59,7 @@ class PostController extends AbstractController
     #[Route('/post/edit/{id<\d+>}', name: 'edit-post')]
     public function update(Post $post, Request $request, ManagerRegistry $doctrine): Response
     {
+        $this->denyAccessUnlessGranted('IS_AUTHENTICATED__FULLY');
         $form = $this->createForm(PostType::class, $post);
         $form->handleRequest($request);
         if ($form->isSubmitted() && $form->isValid()) {
@@ -69,6 +76,7 @@ class PostController extends AbstractController
     #[Route('/post/copy/{id<\d+>}', name: "copy-post")]
     public function duplicate(Post $post, ManagerRegistry $doctrine): Response
     {
+        $this->denyAccessUnlessGranted('IS_AUTHENTICATED__FULLY');
         $copyPost = clone $post;
         $em = $doctrine->getManager();
         $em->persist($copyPost);
